% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_agents.R
\name{create_item}
\alias{create_item}
\title{Function for creating item objects in simulation}
\usage{
create_item(id, true_ability, starting_value, urn_size)
}
\description{
Function for creating item objects in simulation
}
