% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_agents.R
\name{create_player}
\alias{create_player}
\title{Function for creating player objects in simulation}
\usage{
create_player(id, true_ability, starting_value, urn_size)
}
\description{
Function for creating player objects in simulation
}
