% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urnings_game.R
\name{urnings_game}
\alias{urnings_game}
\title{Function setting up the analysis environment}
\usage{
urnings_game(
  scores,
  urn_size_player = 16,
  urn_size_item = 32,
  init_value_player = 0.5,
  init_value_items = 0.5,
  alg_type = "Urnings2",
  paired_update_bool = TRUE
)
}
\description{
Function setting up the analysis environment
}
