% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urnings_game_sim.R
\name{urnings_game_sim}
\alias{urnings_game_sim}
\title{Assemble the simulation environment for unidimensional Urnings algorithm.}
\usage{
urnings_game_sim(
  players,
  items,
  n_games,
  test = TRUE,
  alg_type,
  adaptivity,
  paired_update_bool,
  mu = 0,
  sigma = 2
)
}
\description{
Assemble the simulation environment for unidimensional Urnings algorithm.
}
