% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urningsR.R
\docType{package}
\name{urningsR}
\alias{urningsR}
\title{urningsR: Simulation and data analysis using algorithms for adaptive learning systems}
\description{
This package contains function with which one can simulate the behaviour of
adaptive learning systems, and can analyse exsisting datasets coming from
adaptive learning systesms. The focus of the package is the Urnings algorithm,
and its variants for response time and multidimensional data, but it also
includes wrappers for the Elo rating system.
}
\section{Simulation methods}{


\emph{urnings_game_sim}: sets the simulation environment for unidimensional Urnings

\emph{urnings_game_rt_sim}: sets the simulation environment for Urnings with response time

\emph{mv_urnings_game_sim}:sets the simulation environment for multidimensional Urnings

\emph{play}: executes the simulation
}

\section{Data analysis methods}{


\emph{urnings_game}: sets the environment for unidimensional Urnings

\emph{urnings_game_rt}: sets the environment for Urnings with response time

\emph{mv_urnings_game}:sets the environment for multidimensional Urnings

\emph{analyse}: executes the the analysis
}

